Rename inline_headers_handles{.h,.cc} -> cache_custom_headers{.h,.cc}
