+++ source/extensions/filters/http/cache/cache_filter.cc
@@ -7,7 +7,7 @@
 #include "common/http/utility.h"
 
 #include "extensions/filters/http/cache/cacheability_utils.h"
-#include "extensions/filters/http/cache/inline_headers_handles.h"
+#include "extensions/filters/http/cache/cache_custom_headers.h"
 
 #include "absl/memory/memory.h"
 #include "absl/strings/str_cat.h"
