+++ source/extensions/filters/http/cache/cacheability_utils.cc
@@ -5,7 +5,7 @@
 #include "common/common/macros.h"
 #include "common/common/utility.h"
 
-#include "extensions/filters/http/cache/inline_headers_handles.h"
+#include "extensions/filters/http/cache/cache_custom_headers.h"
 
 namespace Envoy {
 namespace Extensions {
